#2-2: todo.js, db.js를 main.js에 병합
