--- back/todo.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const app = express();
-
-const todoList = [];
-
-app.get((req, res) => {
-
-});
