+++ package.json
@@ -1,0 +1,13 @@
+{
+  "name": "clone_momentum",
+  "version": "1.0.0",
+  "description": "https://nomadcoders.co/javascript-for-beginners",
+  "main": "main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node main.js"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
